Expanded Math Library & Code cleanup
Hazel::Math
+ Added IsNearlyZero, IsNearlyEqual and IsNearlyNotEqual, wrapper around glm::epsilon methods.

~ Cleaned up more of the code suggested by ReSharper Inspection.
~ Entities can now have circle and box collider at the same time.
~ Replace std::move by const ref.

+++ Hazel/src/Hazel/Renderer/SubTexture2D.cpp
@@ -14,8 +14,8 @@ namespace Hazel
 
 	Ref<SubTexture2D> SubTexture2D::CreateFromCoords(const Ref<Texture2D>& texture, const glm::vec2& coords, const glm::vec2& cellSize, const glm::vec2& spriteSize)
 	{
-		glm::vec2 min = { coords.x * cellSize.x / texture->GetWidth(), coords.y * cellSize.y / texture->GetHeight() };
-		glm::vec2 max = { (coords.x + spriteSize.x) * cellSize.x / texture->GetWidth(),	(coords.y + spriteSize.y) * cellSize.y / texture->GetHeight() };
+		glm::vec2 min = { coords.x * cellSize.x / static_cast<float>(texture->GetWidth()), coords.y * cellSize.y / static_cast<float>(texture->GetHeight()) };
+		glm::vec2 max = { (coords.x + spriteSize.x) * cellSize.x / static_cast<float>(texture->GetWidth()),	(coords.y + spriteSize.y) * cellSize.y / static_cast<float>(texture->GetHeight()) };
 
 		return CreateRef<SubTexture2D>(texture, min, max);
 	}
